+++ euchrecli/util/rule_util.py
@@ -1,21 +1,30 @@
 
-from euchrecli.util.card_util import Card, Suit
-from euchrecli.util.player_util import Player, Team
+from euchrecli.abstract import Player, Team, Card, Suit
 
 
 def valid_play(card_to_play: Card, player_hand: [Card], played_cards: [Card],
                trump_suit: Suit) -> bool:
-    """Return whether a card is a valid play."""
+    """Return whether a card is a valid play.
+
+    Args:
+        card_to_play (Card): Proposed card to play.
+        player_hand ([Card]): List of cards in player's hand.
+        played_cards ([Card]): Cards already played this trick.
+        trump_suit (Suit): Active trump suit.
+
+    Returns:
+        bool: Whether or not the card is a valid play.
+    """
     valid = False
-    reason = 0
+    # reason = 0
     if len(played_cards) == 0 or len(player_hand) == 1:
         valid = True
-        reason = 1
+        # reason = 1
     else:
         lead_suit = played_cards[0].adjusted_suit(trump_suit)
         if card_to_play.adjusted_suit(trump_suit) == lead_suit:
             valid = True
-            reason = 2
+            # reason = 2
         else:
             lead_matches = 0
             for card in player_hand:
@@ -24,14 +33,24 @@ def valid_play(card_to_play: Card, player_hand: [Card], played_cards: [Card],
                     lead_matches += 1
 
             valid = lead_matches == 0
-            reason = 3
+            # reason = 3
 
-    print(f'\t[DEBUG] {card_to_play}, valid ({reason}): {valid}')
+    # print(f'\t[DEBUG] {card_to_play}, valid ({reason}): {valid}')
     return valid
 
 
-def trick_winner(players: [Player], played_cards: [Card], trump_suit: Suit):
-    """Return winning player of trick. Increment their team's trick score."""
+def trick_winner(players: [Player], played_cards: [Card], trump_suit: Suit) \
+        -> Player:
+    """Determine and award winning player and team of trick.
+
+    Args:
+        players ([Player]): List of players.
+        played_cards ([type]): List of cards played in the trick.
+        trump_suit (Suit): Active trump suit.
+
+    Returns:
+        Player: Player that won the trick.
+    """
     lead_suit = played_cards[0].suit
 
     high_value = 0
@@ -44,11 +63,18 @@ def trick_winner(players: [Player], played_cards: [Card], trump_suit: Suit):
 
     # increment team tricks won
     winning_player.won_trick()
-    print(f'Trick winner: {winning_player.name}, {winning_player.team}')
+    return winning_player
 
 
 def hand_winner(teams: [Team]) -> Team:
-    """Return winning team of the last hand. Increment team game_score."""
+    """Return winning team of the last hand. Increment team game_score.
+
+    Args:
+        teams ([Team]): List of teams in the game.
+
+    Returns:
+        Team: Team that won the hand.
+    """
     winning_team = None
     for team in teams:
         if team.trick_score >= 3: